requirements.txt
fastapi
uvicorn[standard]
twilio
google-api-python-client
google-auth
google-auth-httplib2
openai
requests